# Add FlashInfer allreduce RMSNorm Quant fusion