--- a/vllm/compilation/collective_fusion.py
+++ b/vllm/compilation/collective_fusion.py
@@ -361,6 +361,7 @@ class AllReduceRMSNormPattern(BasePattern):
                 residual=residual,
                 norm_out=rms_result,
                 quant_out=None,
+                scale_out=None,
                 rms_gamma=weight,
                 rms_eps=self.epsilon,
                 pattern_code=flashinfer_comm.AllReduceFusionPattern.
@@ -420,6 +421,7 @@ class AllReduceFusedAddRMSNormPattern(BasePattern):
                 residual=residual,
                 norm_out=None,
                 quant_out=None,
+                scale_out=None,
                 rms_gamma=weight,
                 rms_eps=self.epsilon,
                 pattern_code=flashinfer_comm.AllReduceFusionPattern.
@@ -490,6 +492,7 @@ class AllReduceFusedRMSNormStaticQuantFP8Pattern(BasePattern):
                 residual=residual,
                 norm_out=result_rms,
                 quant_out=quant_result,
+                scale_out=None,
                 rms_gamma=weight,
                 rms_eps=self.epsilon,
                 pattern_code=flashinfer_comm.AllReduceFusionPattern.
@@ -572,6 +575,7 @@ class AllReduceFusedAddRMSNormStaticQuantFP8Pattern(BasePattern):
                 residual=residual,
                 norm_out=None,
                 quant_out=quant_result,
+                scale_out=None,
                 rms_gamma=weight,
                 rms_eps=self.epsilon,
                 pattern_code=flashinfer_comm.AllReduceFusionPattern.