## Purpose

Add integration of FlashInfer fused allreduce RMSNorm and Quantization. Enabled by `-O '{"pass_config": {"enable_fi_allreduce_fusion": true}, "level":3, "custom_ops":["+quant_fp8","+rms_norm"]}'`

## Validation
`lm_eval --model local-completions --model_args model=${model},base_url=http://localhost:8000/v1/completions --batch_size auto --trust_remote_code  --cache_requests true --tasks gsm8k --num_fewshot 5 --batch_size 200`
### FP8
model RedHatAI/Meta-Llama-3-8B-Instruct-FP8 TP=2
Baseline:
```
|Tasks|Version|     Filter     |n-shot|  Metric   |   |Value |   |Stderr|
|-----|------:|----------------|-----:|-----------|---|-----:|---|-----:|
|gsm8k|      3|flexible-extract|     5|exact_match|↑  |0.7460|±  | 0.012|
|     |       |strict-match    |     5|exact_match|↑  |0.7483|±  | 0.012|
```
PR:
```
|Tasks|Version|     Filter     |n-shot|  Metric   |   |Value |   |Stderr|                                                                                                                                                                                                             
|-----|------:|----------------|-----:|-----------|---|-----:|---|-----:|
|gsm8k|      3|flexible-extract|     5|exact_match|↑  |0.7513|±  |0.0119|                                                                  
|     |       |strict-match    |     5|exact_match|↑  |0.7528|±  |0.0119|
```
### FP4

model: RedHatAI/Qwen3-32B-NVFP4 TP=2

Baseline
```
|Tasks|Version|     Filter     |n-shot|  Metric   |   |Value |   |Stderr|
|-----|------|----------------|-----|-----------|---|-----|---|-----|
|gsm8k|      3|flexible-extract|     5|exact_match|↑  |0.6634|±  | 0.013|
|     |       |strict-match    |     5|exact_match|↑  |0.7468|±  | 0.012|
```

PR:
```
|Tasks|Version|     Filter     |n-shot|  Metric   |   |Value |   |Stderr|
|-----|------:|----------------|-----:|-----------|---|-----:|---|-----:|
|gsm8k|      3|flexible-extract|     5|exact_match|↑  |0.6687|±  |0.0130|
|     |       |strict-match    |     5|exact_match|↑  |0.7559|±  |0.0118|
```
## Benchmarks
Client: 
```
vllm bench serve \
        --model "$MODEL" \
        --dataset-name sonnet \
        --dataset-path benchmarks/sonnet.txt \
        --request-rate "$qps" \
        --num-prompts $((DURATION_SECONDS * qps)) \
```
Input len: 550, Output len: 150. B200 GPUs

### FP8
RedHatAI/Meta-Llama-3-8B-Instruct-FP8 TP=2

Baseline:
| QPS | Mean TTFT (ms) | Median TTFT (ms) | Mean TPOT (ms) | Median TPOT (ms) | 
|-----|----------------|------------------|----------------|------------------|
| 1   | 14.987         | 14.301           | 3.944          | 3.852            |
| 5   | 16.165         | 16.135           | 4.302          | 4.305            |                   
| 10  | 17.422         | 17.458           | 4.543          | 4.537            |

Fused Custom ops `-O '{"pass_config": {"enable_fi_allreduce_fusion": false, "enable_noop": true, "enable_fusion": true}, "level":3, "custom_ops":["+quant_fp8","+rms_norm"]}'`:
| QPS | Mean TTFT (ms) | Median TTFT (ms) | Mean TPOT (ms) | Median TPOT (ms) | 
|-----|----------------|------------------|----------------|------------------|
| 1   | 15.084         | 14.685           | 4.150          | 4.094            |
| 5   | 17.001         | 17.048           | 4.533          | 4.531            |
| 10  | 18.065         | 17.950           | 4.781          | 4.780            |

Non fused Custom ops `-O '{"pass_config": {"enable_fi_allreduce_fusion": false, "enable_noop": false, "enable_fusion": false}, "level":3, "custom_ops":["+quant_fp8","+rms_norm"]}'`:

| QPS | Mean TTFT (ms) | Median TTFT (ms) | Mean TPOT (ms) | Median TPOT (ms) | 
|-----|----------------|------------------|----------------|------------------|
| 1   | 15.233         | 14.456           | 4.249          | 4.175            |
| 5   | 17.266         | 17.147           | 4.634          | 4.634            |
| 10  | 18.457         | 18.254           | 4.904          | 4.902            |

PR:
`-O '{"pass_config": {"enable_fi_allreduce_fusion": true, "enable_noop": false, "enable_fusion": false}, "level":3, "custom_ops":["+quant_fp8","+rms_norm"]}'`
| QPS | Mean TTFT (ms) | Median TTFT (ms) | Mean TPOT (ms) | Median TPOT (ms) | 
|-----|----------------|------------------|----------------|------------------|
| 1   | 14.807         | 14.384           | 3.843          | 3.753            |
| 5   | 16.479         | 16.508           | 4.232          | 4.238            |
| 10  | 17.307         | 17.100           | 4.464          | 4.457            |

### NVFP4
model: RedHatAI/Qwen3-32B-NVFP4 TP=2
Baseline:
| QPS | Mean TTFT (ms) | Median TTFT (ms) | Mean TPOT (ms) | Median TPOT (ms) | 
|-----|----------------|------------------|----------------|------------------|
| 1   | 37.002         | 35.057           | 10.754         | 10.824           |
| 5   | 39.015         | 38.259           | 12.020         | 11.960           |      
| 10  | 44.029         | 42.588           | 14.114         | 14.170           |

Non Fused Custom Ops:
`-O '{"pass_config": {"enable_fi_allreduce_fusion": false, "enable_noop": false, "enable_fusion": false}, "level":3, "custom_ops":["+rms_norm"]}'`

| QPS | Mean TTFT (ms) | Median TTFT (ms) | Mean TPOT (ms) | Median TPOT (ms) | 
|-----|----------------|------------------|----------------|------------------|
| 1   | 36.409         | 35.648           | 11.025         | 11.102           |
| 5   | 40.440         | 39.751           | 12.404         | 12.353           |
| 10  | 46.272         | 44.590           | 14.689         | 14.761           |

Fused Custom Ops:
`-O '{"pass_config": {"enable_fi_allreduce_fusion": false, "enable_noop": true, "enable_fusion": true}, "level":3, "custom_ops":["+rms_norm"]}'`
| QPS | Mean TTFT (ms) | Median TTFT (ms) | Mean TPOT (ms) | Median TPOT (ms) | 
|-----|----------------|------------------|----------------|------------------|
| 1   | 36.458         | 36.017           | 11.077         | 11.155           |
| 5   | 40.235         | 39.002           | 12.440         | 12.377           |
| 10  | 46.241         | 44.097           | 14.758         | 14.789           |

PR:
`-O '{"pass_config": {"enable_fi_allreduce_fusion": true, "enable_noop": false, "enable_fusion": false}, "level":3, "custom_ops":["+rms_norm"]}'`

| QPS | Mean TTFT (ms) | Median TTFT (ms) | Mean TPOT (ms) | Median TPOT (ms) | 
|-----|----------------|------------------|----------------|------------------|
| 1   | 34.068         | 33.255           | 10.346         | 10.449           |
| 5   | 41.480         | 39.513           | 11.673         | 11.648           |
| 10  | 44.979         | 42.499           | 13.400         | 13.411           |

In the case of FP8 quantization PR improves TPOT compared to Custom ops by ~7-8%, no speedup compared torch.compiled (default) ops. Increases TTFT by up to 8%.
In case of NVFP4 PR improves TPOT by up to 5% compared to torch.compiled ops,  TTFT by up to 5%. Comparing to custom ops, PR improves TPOT by up to 10-15%, TTFT by up to 7%.

## Profiling
model: RedHatAI/Qwen3-32B-NVFP4 TP=2


Baseline:
<img width="1088" height="391" alt="image" src="https://github.com/user-attachments/assets/98f5bb78-d50e-4333-aa0b-b43b1799cbea" />

PR:
<img width="582" height="314" alt="image" src="https://github.com/user-attachments/assets/7d72fef9-34fc-4ead-8e5e-f758d11f9d0e" />