@mgoin 2025-07-31
@ilmarkov have you been able to reproduce my FP4 failure?