@ilmarkov 2025-07-31
@mgoin No, it worked on sm100 and I disabled the test on SM < 100