@ilmarkov 2025-08-12
@weireweire I am working on the thresholds in this [PR](22086). We need to divide by tp_size because of one_shot initialization in FlashInfer. It allocates this buffer for each peer rank so we need to make sure that it doesn't fail.