@weireweire 2025-08-13
@ilmarkov thanks, if so can we just change
```
        use_flashinfer = current_tensor_size <= min(
            _FI_MAX_SIZES.get(world_size, _DEFAULT_FI_MAX_SIZE),
            max_fusion_size,
        )
```
to 
```
        use_flashinfer = (current_tensor_size <= max_fusion_size)
```
otherwise it seems redundant.