@weireweire 2025-08-14
could you add a warning [HERE](https://github.com/vllm-project/vllm/blob/586f286789a09f5616be74ee8bedde0a9f698a72/vllm/config.py#L4162) to make user enable noop pass like other backend, otherwise fusion won't happen.